Add files via upload
I found a website on the dark net (the onion router) and it was interesting with mathematical stuff so I thought it was about ancient science, especially when they put the X card from Hunter X Hunter which I thought was cool. but I recognized the language it was written in (the Tifinagh, used in North Africa) and made this thing to translate it. they were asking for bitcoin.